requirements.txt
django
django-rest-framework
numpy
pandas
tensorflow
scikit-learn